Refactor the code related to NOUS E6
